refactor(null): rename nullx package to null
No stdlib package is named null, so the x suffix was decoration (unlike
iox/hashx/randx/stringsx/subtlex, which dodge io/hash/rand/strings/subtle).
No external consumers; consistent with the existing set.Set.
