+++ null/null_test.go
@@ -1,26 +1,26 @@
-package nullx_test
+package null_test
 
 import (
 	"encoding/json"
 	"testing"
 
-	"github.com/dmitrymomot/forge/nullx"
+	"github.com/dmitrymomot/forge/null"
 )
 
 func TestOfEmptyGet(t *testing.T) {
-	if v, ok := nullx.Of("hi").Get(); !ok || v != "hi" {
+	if v, ok := null.Of("hi").Get(); !ok || v != "hi" {
 		t.Fatalf("got %q ok=%v", v, ok)
 	}
-	if _, ok := nullx.Empty[string]().Get(); ok {
+	if _, ok := null.Empty[string]().Get(); ok {
 		t.Fatal("want invalid")
 	}
 }
 
 func TestPtrAndFromPtr(t *testing.T) {
-	if nullx.Empty[int]().Ptr() != nil {
+	if null.Empty[int]().Ptr() != nil {
 		t.Fatal("want nil Ptr")
 	}
-	orig := nullx.Of(9)
+	orig := null.Of(9)
 	p := orig.Ptr()
 	if p == nil || *p != 9 {
 		t.Fatalf("got %v", p)
@@ -31,25 +31,25 @@ func TestPtrAndFromPtr(t *testing.T) {
 	}
 
 	x := 3
-	if v, ok := nullx.FromPtr(&x).Get(); !ok || v != 3 {
+	if v, ok := null.FromPtr(&x).Get(); !ok || v != 3 {
 		t.Fatalf("got %d ok=%v", v, ok)
 	}
-	if _, ok := nullx.FromPtr[int](nil).Get(); ok {
+	if _, ok := null.FromPtr[int](nil).Get(); ok {
 		t.Fatal("want invalid for nil pointer")
 	}
 }
 
 func TestJSON(t *testing.T) {
-	b, err := json.Marshal(nullx.Of(5))
+	b, err := json.Marshal(null.Of(5))
 	if err != nil || string(b) != "5" {
 		t.Fatalf("got %s, %v", b, err)
 	}
-	b, err = json.Marshal(nullx.Empty[int]())
+	b, err = json.Marshal(null.Empty[int]())
 	if err != nil || string(b) != "null" {
 		t.Fatalf("got %s, %v", b, err)
 	}
 
-	var n nullx.Null[int]
+	var n null.Null[int]
 	if err := json.Unmarshal([]byte("null"), &n); err != nil {
 		t.Fatal(err)
 	}
@@ -65,12 +65,12 @@ func TestJSON(t *testing.T) {
 }
 
 func TestSQLRoundTrip(t *testing.T) {
-	n := nullx.Of[string]("hi")
+	n := null.Of[string]("hi")
 	v, err := n.Value()
 	if err != nil {
 		t.Fatal(err)
 	}
-	var m nullx.Null[string]
+	var m null.Null[string]
 	if err := m.Scan(v); err != nil {
 		t.Fatal(err)
 	}
@@ -78,7 +78,7 @@ func TestSQLRoundTrip(t *testing.T) {
 		t.Fatalf("got %q ok=%v", got, ok)
 	}
 
-	e := nullx.Empty[int64]()
+	e := null.Empty[int64]()
 	ev, err := e.Value()
 	if err != nil {
 		t.Fatal(err)
@@ -86,7 +86,7 @@ func TestSQLRoundTrip(t *testing.T) {
 	if ev != nil {
 		t.Fatalf("want nil driver value, got %v", ev)
 	}
-	var me nullx.Null[int64]
+	var me null.Null[int64]
 	if err := me.Scan(nil); err != nil {
 		t.Fatal(err)
 	}
